collect to_string tests
